backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo<4.7
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
pytest-asyncio
httpx
mongomock
